+++ src/Resources/contao/languages/de/tl_settings.php
@@ -3,43 +3,13 @@
 /**
  * legends
  */
-$GLOBALS['TL_LANG']['tl_settings']['wertungsportal_legend']       = 'DeWIS';
+$GLOBALS['TL_LANG']['tl_settings']['wertungsportal_legend']       = 'Wertungsportal';
 
 /**
  * fields
  */
-$GLOBALS['TL_LANG']['tl_settings']['wertungsportal_switchedOff'] = array('DWZ-Abfrage deaktivieren','Abfragen der DWZ-Daten bei DeWIS deaktivieren');
-$GLOBALS['TL_LANG']['tl_settings']['wertungsportal_switchedOffText'] = array('Grund der Abschaltung','Der Text wird im Frontend angezeigt.');
-
-$GLOBALS['TL_LANG']['tl_settings']['wertungsportal_cache'] = array('Cache aktivieren','Cache für DeWIS aktivieren');
-$GLOBALS['TL_LANG']['tl_settings']['wertungsportal_cache_default'] = array('Speicherzeit Standard','Speicherzeit des Standardcaches in Stunden');
-$GLOBALS['TL_LANG']['tl_settings']['wertungsportal_cache_verband'] = array('Speicherzeit Verbände/Vereine','Speicherzeit des Caches für Verbände/Vereine in Stunden');
-$GLOBALS['TL_LANG']['tl_settings']['wertungsportal_cache_referent'] = array('Speicherzeit Referenten','Speicherzeit des Caches für Referenten in Stunden');
-
-$GLOBALS['TL_LANG']['tl_settings']['wertungsportal_karteisperre_gaeste'] = array('Karteikarten für Gäste sperren','Nichtangemeldeten Besuchern den Zugriff auf die Karteikarten verweigern.');
-$GLOBALS['TL_LANG']['tl_settings']['wertungsportal_passive_ausblenden'] = array('Passiv-Mitgliedschaften ausblenden','Passiv gemeldete Spieler ausblenden');
-$GLOBALS['TL_LANG']['tl_settings']['wertungsportal_geburtsjahr_ausblenden'] = array('Geburtsjahr ausblenden','Geburtsjahr der Spieler ausblenden');
-$GLOBALS['TL_LANG']['tl_settings']['wertungsportal_geschlecht_ausblenden'] = array('Geschlecht ausblenden','Geschlecht der Spieler ausblenden');
-
-$GLOBALS['TL_LANG']['tl_settings']['wertungsportal_seite_spieler'] = array('Spielerseite wählen','Seite, auf der das Spieler-Modul eingebunden ist');
-$GLOBALS['TL_LANG']['tl_settings']['wertungsportal_seite_turnier'] = array('Turnierseite wählen','Seite, auf der das Turnier-Modul eingebunden ist');
-$GLOBALS['TL_LANG']['tl_settings']['wertungsportal_seite_verein'] = array('Vereinseite wählen','Seite, auf der das Verein-Modul eingebunden ist');
-$GLOBALS['TL_LANG']['tl_settings']['wertungsportal_seite_verband'] = array('Verbandseite wählen','Seite, auf der das Verband-Modul eingebunden ist');
-
-$GLOBALS['TL_LANG']['tl_settings']['wertungsportal_elobase'] = array('EloBase-Datenbank aktivieren','Zugriff auf alte EloBase-Datenbank aktivieren');
-$GLOBALS['TL_LANG']['tl_settings']['wertungsportal_elobase_host'] = array('Host', 'Host-Adresse');
-$GLOBALS['TL_LANG']['tl_settings']['wertungsportal_elobase_db'] = array('Datenbank', 'Datenbank');
-$GLOBALS['TL_LANG']['tl_settings']['wertungsportal_elobase_user'] = array('Benutzer', 'Benutzer');
-$GLOBALS['TL_LANG']['tl_settings']['wertungsportal_elobase_pass'] = array('Passwort', 'Passwort');
-$GLOBALS['TL_LANG']['tl_settings']['wertungsportal_elobase_url'] = array('Link', 'Link zur alten Karteikarte');
-
-$GLOBALS['TL_LANG']['tl_settings']['wertungsportal_playerDefaultImage'] = array('Spielerbild', 'Standardbild für Spieler');
-$GLOBALS['TL_LANG']['tl_settings']['wertungsportal_playerImageSize'] = array('Bildgröße', 'Größe der Spielerbilder');
-$GLOBALS['TL_LANG']['tl_settings']['wertungsportal_clubDefaultImage'] = array('Vereinsbild', 'Standardbild/-logo für Vereine');
-$GLOBALS['TL_LANG']['tl_settings']['wertungsportal_clubImageSize'] = array('Bildgröße', 'Größe der Vereinsbilder/Logos');
-
-$GLOBALS['TL_LANG']['tl_settings']['wertungsportal_eloLocal'] = array('FIDE-Daten von lokaler Quelle laden', 'Die FIDE-Daten (Nation, Titel, Elo) werden aus der Tabelle tl_dwz_elo geladen und nicht aus DeWIS.');
-
-$GLOBALS['TL_LANG']['tl_settings']['wertungsportal_adminName'] = array('Admin-Name', 'Name des Admins');
-$GLOBALS['TL_LANG']['tl_settings']['wertungsportal_adminMail'] = array('Admin-E-Mail', 'E-Mail des Admins');
-$GLOBALS['TL_LANG']['tl_settings']['wertungsportal_apiSubject'] = array('Betreff für die API-E-Mail', 'Betreff für die API-E-Mail');
+$GLOBALS['TL_LANG']['tl_settings']['wertungsportal_apiBasisURL'] = array('API-Basisadresse', 'URL der API-Basis (ohne endenden Slash)');
+$GLOBALS['TL_LANG']['tl_settings']['wertungsportal_tokenURL'] = array('Token-Adresse', 'URL des Token-Endpoints (ohne endenden Slash)');
+$GLOBALS['TL_LANG']['tl_settings']['wertungsportal_clientID'] = array('Client-ID', 'Client-ID');
+$GLOBALS['TL_LANG']['tl_settings']['wertungsportal_clientSecret'] = array('Client Secret', 'Client Secret');
+$GLOBALS['TL_LANG']['tl_settings']['wertungsportal_scopeListe'] = array('Scope DWZ-Liste', 'Scope DWZ-Liste');