+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Bridge Wertungsportal -> DeWIS
 
+## Version 0.0.2 (2026-06-01)
+
+* Add: System-Einstellungen
+* Add: Klasse Wertungsportal für den Zugriff auf die neue API
+* Add: api.php -> öffentlicher Aufruf der API für Testzwecke
+
 ## Version 0.0.1 (2026-05-29)
 
 * Erste Alphaversion